% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurevisions.R
\name{show_palette}
\alias{show_palette}
\title{Display colors from one palette}
\usage{
show_palette(palette)
}
\arguments{
\item{palette}{A string with the name of a palette.}
}
\value{
A ggplot object.
}
\description{
This function displays a simple plot showing the colors and hexideciimal codes for one palette.
}
\details{
Current options are "venus", "earth", "mars", "jupiter",
"ceres", "enceladus", "europa", "titan",
"cancri", "hd", "kepler186", "kepler16b", "pegasi", "pso", "trappest",
grand_tour", "atomic_clock", "atomic_red", "atomic_blue", and "atomic_orange".
}
\examples{
show_palette("mars")

}
