% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurevisions.R
\name{futurevisions}
\alias{futurevisions}
\title{Access Visions of the Future color palettes}
\usage{
futurevisions(palette)
}
\arguments{
\item{palette}{a string.}
}
\value{
A vector of colors in hexidecimal formant.
}
\description{
This function returns one of several color palettes based on the Visios of the Future
poster series developed by NASA's Jet Propulsion Laboratory at the California Institute
of Technology. For more information, including links free to full-sized versions of the
posters, please see https://www.jpl.nasa.gov/visions-of-the-future/
}
\details{
Current options are "venus", "earth", "mars", "jupiter",
"ceres", "enceladus", "europa", "titan",
"cancri", "hd", "kepler186", "kepler 16b", "pegasi", "pso", "trappest",
grand_tour", "atomic_clock", "atomic_red", "atomic_blue", and "atomic_orange".
}
\examples{
futurevisions("mars")
futurevisions("pso")

require(ggplot2)
ggplot(mpg, aes(cty, hwy, color = factor(cyl))) +
   geom_jitter() +
   scale_color_manual(values = futurevisions("mars"))
}
